add js interop
