+++ lib/src/feedback_widget_response.dart
@@ -1,0 +1,8 @@
+import 'countly_presentable_feedback.dart';
+
+class FeedbackWidgetsResponse {
+  FeedbackWidgetsResponse(this.presentableFeedback, this.error);
+
+  final String? error;
+  final List<CountlyPresentableFeedback> presentableFeedback;
+}
