+++ lib/src/log_level.dart
@@ -1,0 +1,1 @@
+enum LogLevel { INFO, DEBUG, VERBOSE, WARNING, ERROR }
